requirements.txt
# general moduls
django
python-decouple
pillow
djangoredstframework

# third party moduls
markdown       # Markdown support for the browsable API.
django-filter
coreapi
drf-yasg[validation]

# deployment moduls

# email third party modules
django-mail-templated

# Refactoring and tests
black
flake8
pytest
pytest-django


